Fixing and updating the oracle command
